Revert "Create View Payment split page (#148)"
This reverts commit 7a16466b020a5f5a62f85daebcc813143a20e578.
