+++ package.json
@@ -11,7 +11,7 @@
     "lint:fix": "eslint \"{src/components,src/app}/*/\" --ext=.ts,.tsx --fix"
   },
   "dependencies": {
-    "@0xsplits/splits-sdk-react": "2.0.3-beta.0",
+    "@0xsplits/splits-sdk-react": "^1.3.3",
     "@ethereum-attestation-service/eas-sdk": "1.5.0",
     "@mdx-js/loader": "^3.0.0",
     "@mdx-js/react": "^3.0.0",
@@ -41,7 +41,6 @@
     "@tanstack/react-table": "^8.11.8",
     "@types/mdx": "^2.0.11",
     "@vercel/analytics": "^1.2.2",
-    "alchemy-sdk": "^3.3.1",
     "axios": "^1.6.7",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
