--- src/app/(dashboard)/utils/stringUtils.ts
@@ -1,7 +1,0 @@
-export function truncateString(str: string, num: number) {
-  if (str.length <= num * 2) {
-    return str;
-  } else {
-    return str.slice(0, num) + "..." + str.slice(-num);
-  }
-}
