+++ src/app/api/payments/service/paymentAddressService.ts
@@ -1,57 +1,19 @@
 import prisma from "db";
-import { PaymentAddressDto } from "../route";
 
-export async function createPaymentAddress(
-  userId: string,
-  teamId: string,
-  paymentAddress: PaymentAddressDto,
-): Promise<PaymentAddressDto | undefined> {
-  try {
-    const foundTeam = await prisma.team.findFirst({
-      where: {
-        id: teamId,
-        owner_user_id: userId,
-      },
-    });
-    if (foundTeam == null) {
-      console.error(
-        `Team was not found for user userId:[${userId}] teanId:[${teamId}]`,
-      );
-      return;
-    }
+export type PaymentAddressDto = {
+  id: string;
+  chain_id: string;
+  team_id: string;
+  wallet_address: string;
+  created_at: Date;
+  payment_receipents: PaymentRecipientDto[];
+};
 
-    const createdPaymentAddress = await prisma.paymentAddress.create({
-      data: {
-        chain_id: paymentAddress.chain_id,
-        wallet_address: paymentAddress.wallet_address,
-        team_id: teamId,
-      },
-    });
-
-    if (!createdPaymentAddress) {
-      return;
-    } else {
-      await prisma.paymentRecipient.createMany({
-        data: paymentAddress.payment_receipents.map((receipent) => ({
-          ...receipent,
-          payment_address_id: createdPaymentAddress.id,
-        })),
-      });
-      const paymentRecipients = await prisma.paymentRecipient.findMany({
-        where: {
-          payment_address_id: createdPaymentAddress.id,
-        },
-      });
-      return {
-        ...createdPaymentAddress,
-        payment_receipents: paymentRecipients ? paymentRecipients : [],
-      };
-    }
-  } catch (error) {
-    console.error(error);
-    return;
-  }
-}
+export type PaymentRecipientDto = {
+  id: string;
+  wallet_address: string;
+  payment_percentage: number;
+};
 
 export async function fetchTeamPaymentAddresses(
   teamId: string,
@@ -76,7 +38,7 @@ export async function fetchTeamPaymentAddresses(
       };
     }
   } catch (error) {
-    console.error(error);
+    console.log(error);
     return;
   }
 }
