+++ src/components/payments/contributors/paymentSplitsTable.tsx
@@ -5,16 +5,13 @@ import { useEffect, useState } from "react";
 import { PaymentSplitsDataTable } from "./paymentSplitsDataTable";
 import { PaymentSplitsColumns } from "./paymentSplitsColumns";
 import { Heading } from "@/components/ui/heading";
-import { PaymentAddressDto } from "@/app/api/payments/route";
 
 interface ProjectContributionTableProps {
   projectId: string;
-  onCreate: (param: PaymentAddressDto) => void;
 }
 
 export default function PaymentSplitsTable({
   projectId,
-  onCreate,
 }: ProjectContributionTableProps) {
   const { data: session } = useSession();
   const [contributors, setContributors] = useState<ContributorDto[]>([]);
@@ -49,8 +46,6 @@ export default function PaymentSplitsTable({
         columns={PaymentSplitsColumns}
         data={contributors}
         isLoading={isLoading}
-        onCreate={onCreate}
-        projectId={projectId}
       ></PaymentSplitsDataTable>
     </div>
   );
