+++ src/components/payments/paymentsOnboarding.tsx
@@ -1,16 +1,13 @@
 import { useState } from "react";
 import OnboardingTutorial from "./onboardingTutorial";
 import PaymentSplitsTable from "./contributors/paymentSplitsTable";
-import { PaymentAddressDto } from "@/app/api/payments/route";
 
 interface PaymentsOnboardingProps {
   projectId: string;
-  onCreate: (param: PaymentAddressDto) => void;
 }
 
 export default function PaymentsOnboarding({
   projectId,
-  onCreate,
 }: PaymentsOnboardingProps) {
   const [onboardingTutorialCompleted, setOnboardingTutorialCompleted] =
     useState(false);
@@ -22,10 +19,7 @@ export default function PaymentsOnboarding({
           onTutorialCompleted={setOnboardingTutorialCompleted}
         ></OnboardingTutorial>
       ) : (
-        <PaymentSplitsTable
-          projectId={projectId}
-          onCreate={onCreate}
-        ></PaymentSplitsTable>
+        <PaymentSplitsTable projectId={projectId}></PaymentSplitsTable>
       )}
     </>
   );
